Enable providing the link address part of the route when writing
This can be used, for instance, to specify the slot that the CPU is
mounted in a rack. The previous way only worked if the CPU was in slot
0.

Also added some comments to the segment definition on the write to avoid
going back to the protocol documentation to understand then.